server-vla/pyproject.toml
[project]
name = "snapvla-server-vla"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
    "snapvla",
    "torch>=2.7",
    "torchvision>=0.22",
    "transformers==4.54.1",
    "tokenizers==0.21.4",
    "timm==1.0.26",
    "bitsandbytes==0.49.2",
    "accelerate>=1.0",
    "einops>=0.7",
    "pillow>=10.0",
    "packaging",
    "safetensors",
    "num2words",
    "websockets>=12",
    "msgpack>=1.0",
    "pydantic>=2.0",
    "numpy>=1.24",
]

[tool.uv.sources]
snapvla = { path = "..", editable = true }
torch = [{ index = "pytorch-cu128", marker = "sys_platform == 'win32'" }]
torchvision = [{ index = "pytorch-cu128", marker = "sys_platform == 'win32'" }]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
